requirements.txt
asynch==0.3.1
certifi==2026.2.25
charset-normalizer==3.4.6
ciso8601==2.3.3
clickhouse-driver==0.2.10
clickhouse-sqlalchemy==0.3.2
greenlet==3.3.2
idna==3.11
leb128==1.0.9
lz4==4.4.5
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.32.5
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3